fix: cells width calculation for the new first row
