+++ admin/src/components/TableHead.jsx
@@ -1,4 +1,4 @@
-import { memo, useContext } from 'react';
+import { memo, useContext, useEffect } from 'react';
 import { __ } from '@wordpress/i18n';
 import classNames from 'classnames';
 import {
@@ -12,8 +12,63 @@ import Stack from '@mui/joy/Stack';
 import { ReactComponent as SettingsIcon } from '../assets/images/menu-icon-settings.svg';
 import { TableContext } from './TableComponent';
 
+const getHeaderCellRealWidth = ( cell ) => {
+	let sortButtonWidth = cell.querySelector( 'button' )?.offsetWidth;
+	let labelSpanWidth = cell.querySelector( 'span.column-label' )?.offsetWidth;
+
+	sortButtonWidth = sortButtonWidth ? sortButtonWidth : 0;
+	labelSpanWidth = labelSpanWidth ? labelSpanWidth : 0;
+	return sortButtonWidth + labelSpanWidth;
+};
+
 const TableHead = () => {
-	const { toggleOpenedRowActions, table, resizable, userCustomSettings, closeableRowActions } = useContext( TableContext );
+	const { tableContainerRef, toggleOpenedRowActions, table, resizable, userCustomSettings, closeableRowActions } = useContext( TableContext );
+
+	// set width of columns according to header items width
+	// default width of cells defined in each table is considered as source width which is used if cell header items (sort button and label) doesnt overflow defined width
+	useEffect( () => {
+		// data cells
+		const nodes = tableContainerRef.current?.querySelectorAll( 'table.urlslab-table thead th:not(.editRow)' );
+		const headerCells = nodes ? Object.values( nodes ) : [];
+		// edit cell
+		const editCell = tableContainerRef.current?.querySelector( 'table.urlslab-table thead th.editRow' );
+
+		for ( const c in headerCells ) {
+			const cell = headerCells[ c ];
+			const totalWidth = getHeaderCellRealWidth( cell ) + 16; // count with paddings
+			const defaultWidth = cell.dataset.defaultwidth ? parseInt( cell.dataset.defaultwidth ) : totalWidth;
+			const finalWidth = totalWidth > defaultWidth ? totalWidth : defaultWidth;
+
+			if ( closeableRowActions ) {
+				cell.style.width = `${ finalWidth }px`;
+				// first cell
+				if ( parseInt( c ) === 0 ) {
+					cell.style.width = `calc(${ finalWidth }px + var(--TableCellFirst-paddingLeft) )`;
+				}
+
+				// last data cell
+				if ( parseInt( c ) === headerCells.length - 1 ) {
+					if ( editCell ) {
+						// make width of last data cell bigger of floating toggle button width to make its text always visible
+						cell.style.width = `calc( ${ finalWidth }px + 2 * var(--TableCell-paddingX) + var(--Table-editRowClosedColumnWidth, 0) )`;
+					} else {
+						// edit cell not present, add just right table padding as it's last cell
+						cell.style.width = `calc( ${ finalWidth }px + 2 * var(--TableCell-paddingX) + var(--TableCellLast-paddingRight) )`;
+					}
+				}
+			} else {
+				cell.style.width = `${ finalWidth }px`;
+				// first cell
+				if ( parseInt( c ) === 0 ) {
+					cell.style.width = `calc(${ finalWidth }px + var(--TableCellFirst-paddingLeft) )`;
+				}
+				// last cell
+				if ( parseInt( c ) === headerCells.length - 1 ) {
+					cell.style.width = `calc(${ finalWidth }px + var(--TableCellLast-paddingRight) )`;
+				}
+			}
+		}
+	}, [ closeableRowActions, userCustomSettings.columnVisibility ] );
 
 	return (
 		<thead className="urlslab-table-head">
